# backend/tests/snapshots/openapi_schema.json
{
  "components": {
    "schemas": {
...
      "AdminCreateUserResponse": {
        "description": "Response for POST /admin/users.",
        "properties": {
          "email": {
            "title": "Email",
            "type": "string"
          },
          "plan_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Plan Id"
          },
          "user_id": {
            "title": "User Id",
            "type": "string"
          }
        },
        "required": [
          "user_id",
          "email"
        ],
        "title": "AdminCreateUserResponse",
        "type": "object"
      },
...
      "AdminUsersListResponse": {
        "description": "Response for GET /admin/users.",
        "properties": {
          "limit": {
            "title": "Limit",
            "type": "integer"
          },
          "offset": {
            "title": "Offset",
            "type": "integer"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          },
          "users": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Users",
            "type": "array"
          }
        },
        "required": [
          "users",
          "total",
          "limit",
          "offset"
        ],
        "title": "AdminUsersListResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "API para busca e an\u00e1lise de licita\u00e7\u00f5es de uniformes no PNCP. Permite filtrar oportunidades por estado, valor e keywords, gerando relat\u00f3rios Excel e resumos executivos via IA.",
    "title": "BidIQ Uniformes API",
    "version": "0.2.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/admin/feature-flags/reload": {
      "post": {
        "description": "Reload all feature flags from environment variables (admin only).\n\nSTORY-226 AC16: Clears the feature flag cache, forcing re-read from\nenvironment on next access. Returns current values after reload.",
        "operationId": "reload_feature_flags_endpoint_admin_feature_flags_reload_post",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Reload Feature Flags Endpoint",
        "tags": [
          "admin"
        ]
      }
    },
    "/admin/users": {
      "get": {
        "description": "List all users with profiles and subscription info.",
        "operationId": "list_users_admin_users_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "search",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 100,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Search"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminUsersListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Users",
        "tags": [
          "admin"
        ]
      },
      "post": {
        "description": "Create a new user with optional plan assignment.",
        "operationId": "create_user_admin_users_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminCreateUserResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create User",
        "tags": [
          "admin"
        ]
      }
    },
    "/admin/users/{user_id}": {
      "delete": {
        "description": "Delete a user and all their data.",
        "operationId": "delete_user_admin_users__user_id__delete",
        "parameters": [
          {
            "description": "User UUID to delete",
            "in": "path",
            "name": "user_id",
            "required": true,
            "schema": {
              "description": "User UUID to delete",
              "title": "User Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminDeleteUserResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Delete User",
        "tags": [
          "admin"
        ]
      },
      "put": {
        "description": "Update user profile or plan.",
        "operationId": "update_user_admin_users__user_id__put",
        "parameters": [
          {
            "description": "User UUID to update",
            "in": "path",
            "name": "user_id",
            "required": true,
            "schema": {
              "description": "User UUID to update",
              "title": "User Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminUpdateUserResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Update User",
        "tags": [
          "admin"
        ]
      }
    },
    "/admin/users/{user_id}/assign-plan": {
      "post": {
        "description": "Manually assign a plan to a user (bypasses payment).",
        "operationId": "assign_plan_admin_users__user_id__assign_plan_post",
        "parameters": [
          {
            "description": "User UUID to assign plan to",
            "in": "path",
            "name": "user_id",
            "required": true,
            "schema": {
              "description": "User UUID to assign plan to",
              "title": "User Id",
              "type": "string"
            }
          },
          {
            "description": "Plan ID to assign",
            "in": "query",
            "name": "plan_id",
            "required": true,
            "schema": {
              "description": "Plan ID to assign",
              "title": "Plan Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminAssignPlanResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Assign Plan",
        "tags": [
          "admin"
        ]
      }
    },
    "/admin/users/{user_id}/credits": {
      "put": {
        "description": "Manually adjust a user's credits (admin only).\n\nSets the credits_remaining value directly on the user's active subscription.\nIf no active subscription exists, creates one based on user's current plan.\n\nArgs:\n    user_id: UUID of the user\n    req.credits: New credit value (must be >= 0)\n\nReturns:\n    Updated user info with new credit value",
        "operationId": "update_user_credits_admin_users__user_id__credits_put",
        "parameters": [
          {
            "description": "User UUID to update credits for",
            "in": "path",
            "name": "user_id",
            "required": true,
            "schema": {
              "description": "User UUID to update credits for",
              "title": "User Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCreditsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminUpdateCreditsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Update User Credits",
        "tags": [
          "admin"
        ]
      }
    },
...
    "/v1/admin/feature-flags/reload": {
      "post": {
        "description": "Reload all feature flags from environment variables (admin only).\n\nSTORY-226 AC16: Clears the feature flag cache, forcing re-read from\nenvironment on next access. Returns current values after reload.",
        "operationId": "reload_feature_flags_endpoint_v1_admin_feature_flags_reload_post",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Reload Feature Flags Endpoint",
        "tags": [
          "admin"
        ]
      }
    },
...
  }
}